Update food-world-cup-data.csv
I have been using Vega-Lite to perform some visualization as a part of my class assignment but as the ratings were present in multiple columns, it was hard for me to perform the visualization. I stacked the ratings of all the countries on top of each other which I feel that would make it easier to visualize.